Add script to automatically find libraries
Libraries required to run the Swift executable are now found automatically using ldd. This avoids the need to store the libraries in the repo.

Because of a mismatch in glibc libraries between smithmicro/swift:3.0 (Ubuntu 15.x) and amazonlinux:2016.09 (CentOS), execution is now via ld-linux.so. This idea came from this article: http://stackoverflow.com/questions/8657908/deploying-yesod-to-heroku-cant-build-statically/8658468#8658468